requirements.txt
# ── Core framework ────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# ── NLP / ML models ──────────────────────────
transformers>=4.40.0
torch>=2.2.0
scikit-learn>=1.4.0

# ── Vision ───────────────────────────────────
Pillow>=10.3.0

# ── Translation & language detection ─────────
deep-translator>=1.11.4
langdetect>=1.0.9

# ── Text preprocessing ───────────────────────
symspellpy>=6.7.7
emoji>=2.10.0

# ── MongoDB ──────────────────────────────────
pymongo>=4.6.0

# ── HTTP ─────────────────────────────────────
requests>=2.31.0

# ── Environment ──────────────────────────────
python-dotenv>=1.0.0
